requirements.txt
# requirements.txt
uvicorn==0.30.1
fastapi==0.111.0
python-dotenv==1.0.1
pypdf==4.2.0
qdrant-client==1.9.0
google-generativeai==0.7.0
watchdog==4.0.1
asyncio==3.4.3